Separate configs and conform with StageConfig
