+++ crates/bevy_ecs/src/system/config/mod.rs
@@ -1,0 +1,13 @@
+mod system_config;
+mod ambiguity_config;
+mod exclusive_config;
+mod run_criteria_config;
+mod schedule_config;
+mod stage_config;
+
+pub use system_config::*;
+pub use ambiguity_config::*;
+pub use exclusive_config::*;
+pub use run_criteria_config::*;
+pub use schedule_config::*;
+pub use stage_config::*;
