+++ crates/bevy_ecs/src/system/config/system_config.rs
@@ -1,0 +1,37 @@
+use crate::{schedule::*, system::InsertionPoint};
+
+#[derive(Default)]
+pub struct SystemConfig {
+    pub labels: Vec<BoxedSystemLabel>,
+    pub before: Vec<BoxedSystemLabel>,
+    pub after: Vec<BoxedSystemLabel>,
+    pub stage: Option<BoxedStageLabel>,
+    pub ambiguity_sets: Vec<BoxedAmbiguitySetLabel>,
+    pub run_criteria: Option<RunCriteriaDescriptorOrLabel>,
+    pub insertion_point: Option<InsertionPoint>,
+}
+
+impl SystemConfig {
+    pub fn add_label(&mut self, label: impl SystemLabel) {
+        self.labels.push(Box::new(label));
+    }
+    pub fn add_before(&mut self, label: impl SystemLabel) {
+        self.before.push(Box::new(label));
+    }
+    pub fn add_after(&mut self, label: impl SystemLabel) {
+        self.after.push(Box::new(label));
+    }
+    pub fn set_stage(&mut self, label: impl StageLabel) {
+        self.stage = Some(Box::new(label));
+    }
+    pub fn add_ambiguity_set(&mut self, set: impl AmbiguitySetLabel) {
+        self.ambiguity_sets.push(Box::new(set));
+    }
+
+    pub fn set_run_criteria(&mut self, criteria: RunCriteriaDescriptorOrLabel) {
+        self.run_criteria = Some(criteria);
+    }
+}
+
+pub struct ParallelSystemKind;
+pub struct SystemSetKind;
